feat(user following): refactor tests
- remove errors in tests
- put authenticated user and new user in setup

[Delivers #162948846]
